Separate device context functions into different files
